will update tomorrow

+++ src/dev/ttetris/Model.java
@@ -32,19 +32,19 @@ public class Model {
         next = new int[4][4][4];
         board = new int[ROW][COL][HIG];
     }
-    /*
+
     // put generated Block into Next area
     public void putNextBlock(Block b) {
         Cube [] cubes = b.getCubes();
         int n = cubes.length;
         for (int i = 0; i < n; i++) {
-            int x = cubes[i].getX();// + b.centerX;
-            int y = cubes[i].getY();// + b.centerY;
-            int z = cubes[i].getZ();// + b.centerZ;
+            int x = (int)cubes[i].getX();// + b.centerX;
+            int y = (int)cubes[i].getY();// + b.centerY;
+            int z = (int)cubes[i].getZ();// + b.centerZ;
             next[x][y][z] = b.getColor();
         }
     }
-    */
+
     /*
     // delete Block in next area
     public void deleteNextBlock(Block c) {
